Move to keyword only arguments to improve readability
New pylint allows you to seperately set a maximum number of total aguments
and a maximum number of positional arguments. This gave us lots of warnings
on the number of functions that had a confusingly high number of arguments.

Rather than just up the limit in pylintrc it seems like good practice to make
many of these arguments keyword only. This make code more clear, and more robust
as we go forward.